move module `files.py` from `bapsflib.lapd._hdf` to `bapsflib._hdf.utils`, rename module `files.py` -> `file.py`, update 'docs/src' accordingly
