+++ bapsflib/_hdf/utils/tests/test_hdfreadcontrol.py
@@ -19,7 +19,7 @@ import unittest as ut
 from bapsflib._hdf.maps import (ConType, FauxHDFBuilder, HDFMap)
 from bapsflib._hdf.maps.controls.templates import HDFMapControlTemplate
 from bapsflib._hdf.maps.controls.waveform import HDFMapControlWaveform
-from bapsflib.lapd._hdf.files import File
+from bapsflib._hdf.utils.file import File
 from numpy.lib import recfunctions as rfn
 from typing import (Any, Dict, List, Tuple)
 from unittest import mock
