+++ bapsflib/_hdf/utils/tests/test_hdfreaddata.py
@@ -15,7 +15,7 @@ import numpy as np
 import unittest as ut
 
 from bapsflib._hdf.maps import (ConType, FauxHDFBuilder)
-from bapsflib.lapd._hdf.files import File
+from bapsflib._hdf.utils.file import File
 
 from ..hdfreaddata import (HDFReadData, condition_shotnum)
